requirements.txt
absl-py==1.0.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cdt==0.5.23
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
chex==0.1.3
click==8.1.3
cloudpickle==2.1.0
cpprb==10.6.4
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dibs-lib==1.0.1
distrax==0.1.2
dm-haiku==0.0.6
dm-tree==0.1.7
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
GPUtil==1.4.0
gym==0.23.1
gym-notices==0.0.6
idna==3.3
igraph==0.9.10
imageio==2.19.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jax==0.3.13
jaxlib==0.3.10+cuda11.cudnn82
jedi==0.18.1
Jinja2==3.1.2
jmp==0.0.2
joblib==1.1.0
jraph==0.0.6.dev0
json5==0.9.8
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
opt-einsum==3.3.0
optax==0.1.2
ott-jax==0.2.6
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pgmpy==0.1.18
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.11
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-igraph==0.9.10
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
skrebate==0.62
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
sumu==0.1.2
tabulate==0.8.9
tensorflow-probability==0.16.0
terminado==0.15.0
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.11.2
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.12.17
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
yarg==0.1.9
zipp==3.8.0
